+++ src/test/java/janggi/domain/position/RowTest.java
@@ -1,12 +1,12 @@
-package janggi.domain;
+package janggi.domain.position;
 
-import janggi.domain.position.Row;
-import org.junit.jupiter.api.DisplayName;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import static org.assertj.core.api.Assertions.*;
 
 class RowTest {
